refactor: move getFullPath to utils.cpp
