+++ include/utils.h
@@ -2,11 +2,11 @@
 # define UTILS_H
 
 # include <iostream>
-# include "Configuration.hpp"
+#include <sys/wait.h>
 
 std::string toLower(const std::string& str);
 bool isDigitsOnly(const std::string& str);
-
+std::string getFullPath(const std::string &file);
 
 
 #endif
