+++ src/poll-loop/handle-poll-events.cpp
@@ -1,6 +1,6 @@
 #include "read-request.h"
 #include "poll-loop.h"
-#include "Cgi.hpp"
+#include "utils.h"
 #include "session.h"
 
 static void setPfdTrackPollOutOnly(struct pollfd& pfd);
